==> notebooks/2.1-data-cleaning.ipynb <==
### Assessing the Readiness of HEI in Malaysia to Accept Generation Alpha

---

# Data Cleaning

**Drop missing values or fill them in with an automated workflow.**

Machine learning algorithms cannot work with missing features.

## Importing the libraries

import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O
import seaborn as sns

We can refer to the Pandas module using the "dot notation" to call its methods. To read our data (which is in the form of csv files), into a so-called DataFrame structure, we use the method read_csv() and pass in each file name as a string: All the datasets are loaded and storeed in pandas dataframe objects.

## Importing the datasets

%%time
info_content_df = pd.read_csv('../data/raw/Info_Content.csv', index_col='ucid')
info_userdata_df = pd.read_csv('../data/raw/Info_UserData.csv', index_col='uuid')
log_problem_df = pd.read_csv('../data/raw/Log_Problem.csv', index_col='upid')

## Handling missing values

The first thing to do when you get a new dataset is take a look at some of it. This lets you see that it all read in correctly and gives an idea of what's going on with the data. In this case, let's see if there are any missing values, which will be reprsented with NaN or None.

# get the number of missing data points per column
missing_values_count = info_content_df.isnull().sum()

# look at the # of missing points in the first ten columns
missing_values_count

# how many total missing values do we have?
total_cells = np.product(info_content_df.shape)
total_missing = missing_values_count.sum()

# percent of data that is missing
percent_missing = (total_missing/total_cells) * 100
print(percent_missing )

# get the number of missing data points per column
missing_values_count = info_userdata_df.isnull().sum()

# look at the # of missing points in the first ten columns
missing_values_count

# how many total missing values do we have?
total_cells = np.product(info_userdata_df.shape)
total_missing = missing_values_count.sum()

# percent of data that is missing
percent_missing = (total_missing/total_cells) * 100
print(percent_missing)

# replace all NA's with unspecified
info_userdata_df = info_userdata_df.fillna('unspecified')

# get the number of missing data points per column
missing_values_count = log_problem_df.isnull().sum()

# look at the # of missing points in the first ten columns
missing_values_count

# how many total missing values do we have?
total_cells = np.product(log_problem_df.shape)
total_missing = missing_values_count.sum()

# percent of data that is missing
percent_missing = (total_missing/total_cells) * 100
print(percent_missing)

# replace all NA's with false
log_problem_df = log_problem_df.fillna(False)

## Exporting Data

The data is in DataFrame form, it can manipulate it by adding a body mass index column. The cleaned data is exported to the interim data folder which stores the intermediate data that has been transformed.

> Check the dataframe before exporting

info_content_df.head()

info_userdata_df.head()

log_problem_df.head()

## Exporting dataset

#info_content_df.to_csv('../data/interim/Info_Content.csv')
#info_userdata_df.to_csv('../data/interim/Info_UserData.csv')
#log_problem_df.to_csv('../data/interim/Log_Problem.csv')

[HDF5](https://www.hdfgroup.org/solutions/hdf5) is a high-performance data management and storage unit. 

%%time

info_content_df.to_hdf('../data/interim/Info_Content.h5',key='info_content')
info_userdata_df.to_hdf('../data/interim/Info_UserData.h5',key='info_userdata')
log_problem_df.to_hdf('../data/interim/Log_Problem.h5',key='log_problem')
