==> notebooks/2.2-data-merging.ipynb <==
# 2.2 - Data Merging

## Importing Libraries

# data processing, CSV file I/O
import pandas as pd

## Importing Data

%%time
info_content_df = pd.read_csv('../data/raw/Info_Content.csv', index_col='ucid')
info_userdata_df = pd.read_csv('../data/raw/Info_UserData.csv', index_col='uuid')
log_problem_df = pd.read_csv('../data/interim/Log_Problem.csv', index_col='upid')

## Merging Data

merge_df = log_problem_df.merge(info_userdata_df,how='left',on ='uuid')

merge_df = merge_df.merge(info_content_df,how='left',on ='ucid')

## Exporting Data

merge_df.to_csv('../data/interim/Cleaned_Log_Problem.csv')
