==> notebooks/2.2-handling-categorical-attributes.ipynb <==
# 2.2 Handling categorical attributes

An error will occur if these categorical varriables are included into most of the machine learning models without preprocessing them first. There are three approach to prepare the categorical data.

- Drop
- Label Encoding
- One-Hot Encoding

## Setting environment

The required packages are imported.

import pandas as pd
import numpy as np

## Importing datasets

Each datasets are imported into different dataframes.

%%time
info_content_df = pd.read_csv('../data/interim/Info_Content.csv', index_col='ucid')
info_userdata_df = pd.read_csv('../data/interim/Info_UserData.csv', index_col='uuid')
log_problem_df = pd.read_csv('../data/interim/Log_Problem.csv', index_col='upid')

info_content_df.head()

info_userdata_df.head()

log_problem_df.head()

## Identifying categorical attributes

s = info_content_df.dtypes == 'object'
object_cols_info_content = list(s[s].index)

print("Categorical variables: ")
print(object_cols_info_content)

s = info_userdata_df.dtypes == 'object'
object_cols_info_userdata = list(s[s].index)

print("Categorical variables: ")
print(object_cols_info_userdata)

s = log_problem_df.dtypes == 'object'
object_cols_log_problem = list(s[s].index)

print("Categorical variables: ")
print(object_cols_log_problem)

### 1. Drop categorical variables

The easiest way to deal with categorical variable is to simply remove the columns from the datasets if the attributes do not contain useful information.

The content_kind and subject attribute only contain one unique value. This would not affect

selected_columns = ['content_pretty_name','difficulty','learning_stage']
info_content_df = info_content_df[selected_columns].copy()

### 2. Label Encoding

**Mapping ordinal features**

To make sure that the learning algorithm interprets the ordinal features correctly, we need to convert the categorical string values into integers. Unfortunately, there is no convenient function that can automatically derive the correct order of the labels of our size feature, so we have to define the mapping manually. In the following simple example, let's assume that we know the numerical difference between features, for example, hard = normal + 1 = easy + 2 = unset + 3.

difficulty_mapping = {
    'easy': 1,
    'normal': 2,
    'hard': 3,
    'unset': 0
}
learning_stage_mapping = {
    'elementary': 1,
    'junior': 2,
    'senior': 3
}

info_content_df['difficulty'] = info_content_df['difficulty'].map(difficulty_mapping)
info_content_df['learning_stage'] = info_content_df['learning_stage'].map(learning_stage_mapping)

info_content_df.head()

info_content_df.describe()

### 3. One-hot encoding - Nominal

By including dummy variable in a regression model however, one should be careful of the Dummy Variable Trap. The Dummy Variable trap is a scenario in which the independent variables are multicollinear - a scenario in which two or more variables are highly correlated; in simple terms one variable can be predicted from the others.

is_downgrade_mapping = {
    True: 1,
    False: 0
}
is_upgrade_mapping = {
    True: 1,
    False: 0
}

log_problem_df['is_downgrade'] = log_problem_df['is_downgrade'].map(is_downgrade_mapping)
log_problem_df['is_upgrade'] = log_problem_df['is_upgrade'].map(is_upgrade_mapping)

log_problem_df.head()

info_userdata_gender = pd.get_dummies(info_userdata_df.gender, prefix='gender').iloc[:,:-1]
info_userdata_df = pd.concat([info_userdata_df, info_userdata_gender], axis=1, sort=False)

info_userdata_city = pd.get_dummies(info_userdata_df.user_city, prefix='user_city').iloc[:,:-1]
info_userdata_df = pd.concat([info_userdata_df, info_userdata_city], axis=1, sort=False)

info_userdata_df = info_userdata_df.drop(columns = ['gender','user_city'])

The info_userdata dataframe can be glanced through head().

info_userdata_df.head()

info_userdata_df.describe()

## Exporting Data

The data is in DataFrame form, it can manipulate it by adding a body mass index column. The transformed data is exported to the interim data folder which stores the intermediate data that has been transformed.

info_content_df.head()

info_userdata_df.head()

log_problem_df.head()

info_content_df.to_csv('../data/interim/Info_Content.csv')
info_userdata_df.to_csv('../data/interim/Info_UserData.csv')
log_problem_df.to_csv('../data/interim/Log_Problem.csv')
