==> notebooks/2.4-data-merging.ipynb <==
# 2.2 - Data Merging

Oftentimes data will come to us with column names, index names, or other naming conventions that we are not satisfied with. In that case, you'll learn how to use pandas functions to change the names of the offending entries to something better.

## Importing Libraries

# data processing, CSV file I/O
import pandas as pd

## Importing Data

%%time
info_content_df = pd.read_csv('../data/interim/Info_Content.csv', index_col='ucid')
info_userdata_df = pd.read_csv('../data/interim/Info_UserData.csv', index_col='uuid')
log_problem_df = pd.read_csv('../data/interim/Log_Problem.csv')

info_content_df.head()

info_userdata_df.head()

log_problem_df.head()

## Merging Data

merge_df = log_problem_df.merge(info_userdata_df,how='left',on='uuid')

merge_df = merge_df.merge(info_content_df,how='left',on='ucid')

merge_df.head()

merge_df.columns

merge_df.shape

## Exporting Data

%%time
merge_df.to_csv('../data/interim/Merged_Data.csv', index=False)
