==> notebooks/2.5-data-spliting.ipynb <==
# Data Spliting

## Setting Environment

import pandas as pd
import numpy as np

## Importing Data

%%time
merge_df = pd.read_csv('../data/processed/Merged_Data.csv')

merge_df.head()

merge_df.shape

merge_df.info()

X = merge_df.drop('is_proficient', axis=1)
y = merge_df['is_proficient']

## Splitting Data

# Splitting the dataset into the Training set and Test set
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.25, random_state = 0)

X_train.shape, X_test.shape, y_train.shape, y_test.shape

print(X_train)
print(y_train)

print(X_test)
print(y_test)

## Export dataframes

%%time
X_train.to_csv('../data/interim/X_train.csv', index=False)
X_test.to_csv('../data/interim/X_test.csv', index=False)
y_train.to_csv('../data/interim/y_train.csv', index=False)
y_test.to_csv('../data/interim/y_test.csv', index=False)
