==> notebooks/2.6-feature-selection.ipynb <==
# Feature Selection

1. Remove irrelevant features results in a better performance model
2. in a
3. in a model that run faster

# Importing libraries

import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

%%time
X_train = pd.read_csv('../data/interim/X_train.csv')
y_train = pd.read_csv('../data/interim/y_train.csv')
X_test = pd.read_csv('../data/interim/X_test.csv')
y_test = pd.read_csv('../data/interim/y_test.csv')

X_train.info()

### Amount of variation

The variable has to be reviewed and dropped if the variable has a very low variation, because the model is not going to learn anything with it.

### Pairwise correlations

a

### PCA

Principal Component Analysis, or PCA, is a dimensionality-reduction method that is often used to reduce the dimensionality of large data sets, by transforming a large set of variables into a smaller one that still contains most of the information in the large set.

Like every other data transformation, we can apply PCA to a dataset in scikit-learn with a transformer, in this case the PCA module. When initializing the PCA module, we can use the n_components keyword to specify the number of principal components. The default setting is to extract m - 1 principal components, where m is the number of features in the dataset.

# Applying PCA
from sklearn.decomposition import PCA
pca = PCA(n_components = 2)
X_train = pca.fit_transform(X_train)
X_test = pca.transform(X_test)

X_train

X_test

## Exporting data

%%time
pd.DataFrame(X_train, columns=['columns']).to_csv("X_train.csv", index = None)
pd.DataFrame(X_test, columns=['columns']).to_csv("X_test.csv", index = None)
y_train.to_csv('../data/interim/y_train.csv', index=False)
y_test.to_csv('../data/interim/y_test.csv', index=False)
