==> notebooks/2.7-feature-scaling.ipynb <==
## Setting Up Environment

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

## Importing Data

%%time
X_train = pd.read_csv('../data/interim/X_train.csv')
y_train = pd.read_csv('../data/interim/y_train.csv')
X_test = pd.read_csv('../data/interim/X_test.csv')
y_test = pd.read_csv('../data/interim/y_test.csv')

## Scaling

sc = StandardScaler()
X_train = sc.fit_transform(X_train)
X_test = sc.transform(X_test)
print(X_train)
print(X_test)

%%time
X_train.to_csv('../data/interim/X_train.csv', index=False)
X_test.to_csv('../data/interim/X_test.csv', index=False)
y_train.to_csv('../data/interim/y_train.csv', index=False)
y_test.to_csv('../data/interim/y_test.csv', index=False)
